Refactor throttle_and_brake with bar_gauge single widget.
