+++ setup.py
@@ -15,7 +15,7 @@ setup(
         ('share/' + package_name + '/resource',
             ['resource/steering_wheel.ui']),
         ('share/' + package_name + '/resource',
-            ['resource/throttle_brake_pedals.ui']),
+            ['resource/bar_gauge.ui']),
         ('share/' + package_name, ['package.xml']),
         ('share/' + package_name, ['plugin.xml']),
     ],
