--- src/rqt_gauges/throttle_brake_pedals_widget.py
@@ -1,147 +1,0 @@
-import os
-
-from ament_index_python.resources import get_resource
-from PyQt5.QtCore import pyqtSignal, pyqtSlot, Qt
-from PyQt5.QtWidgets import QWidget
-from python_qt_binding import loadUi
-from rosidl_runtime_py.utilities import get_message
-from rqt_py_common.topic_completer import TopicCompleter
-
-from .utils import generate_field_evals, get_topic_type
-
-
-class ThrottleBrakePedalsWidget(QWidget):
-
-    updateThrottleValueSignal = pyqtSignal(int, int)
-    updateBrakeValueSignal = pyqtSignal(int, int)
-
-    def __init__(self, node):
-        super().__init__()
-        self.setObjectName('ThrottleBrakePedals_widget')
-
-        self.node = node
-        self.throttle_sub = None
-        self.brake_sub = None
-
-        _, package_path = get_resource('packages', 'rqt_gauges')
-        ui_file = os.path.join(package_path, 'share',
-                               'rqt_gauges', 'resource', 'throttle_brake_pedals.ui')
-        loadUi(ui_file, self)
-
-        # Throttle Topic Completer
-        self.throttle_topic_to_subscribe.setNode(self.node)
-        self._throttle_topic_completer = TopicCompleter(self.throttle_topic_to_subscribe)
-        self._throttle_topic_completer.update_topics(self.node)
-        self.throttle_topic_to_subscribe.setCompleter(self._throttle_topic_completer)
-
-        # Brake Topic Completer
-        self.brake_topic_to_subscribe.setNode(self.node)
-        self._brake_topic_completer = TopicCompleter(self.brake_topic_to_subscribe)
-        self._brake_topic_completer.update_topics(self.node)
-        self.brake_topic_to_subscribe.setCompleter(self._brake_topic_completer)
-
-        # Signals Connection
-        self.throttle_subscribe.pressed.connect(self.throttleUpdateSubscription)
-        self.brake_subscribe.pressed.connect(self.brakeUpdateSubscription)
-        self.updateThrottleValueSignal.connect(self.updateThrottle)
-        self.updateBrakeValueSignal.connect(self.updateBrake)
-
-    def dragEnterEvent(self, event):
-        event.accept()
-
-    def dropEvent(self, event):
-        if event.mimeData().hasText():
-            topic_name = str(event.mimeData().text())
-        else:
-            droped_item = event.source().selectedItems()[0]
-            topic_name = str(droped_item.data(0, Qt.UserRole))
-
-        if (event.pos().x() < self.size().width() / 2):
-            self.throttle_topic_to_subscribe.setText(topic_name)
-            self.throttleUpdateSubscription()
-        else:
-            self.brake_topic_to_subscribe.setText(topic_name)
-            self.brakeUpdateSubscription()
-        event.accept()
-
-    @pyqtSlot()
-    def throttleUpdateSubscription(self):
-        if self.node.destroy_subscription(self.throttle_sub):
-            print('Previous subscription deleted')
-        else:
-            print('There was no previous subscription')
-        topic_path = self.throttle_topic_to_subscribe.text()
-        topic_type, topic_name, fields = get_topic_type(self.node, topic_path)
-        self.throttle_field_evals = generate_field_evals(fields)
-        if topic_type is not None:
-            print('Subscribing to:', topic_name, 'Type:', topic_type, 'Field:', fields)
-            data_class = get_message(topic_type)
-            self.throttle_sub = self.node.create_subscription(
-                data_class,
-                topic_name,
-                self.throttle_callback,
-                10)
-
-    @pyqtSlot(int, int)
-    def updateThrottle(self, value, raw_value):
-        self.throttle_pedal.setValue(value)
-        self.throttle_label.setText(str(raw_value / 100.0))
-
-    @pyqtSlot(int, int)
-    def updateBrake(self, value, raw_value):
-        self.brake_pedal.setValue(value)
-        self.brake_label.setText(str(raw_value / 100.0))
-
-    @pyqtSlot()
-    def brakeUpdateSubscription(self):
-        if self.node.destroy_subscription(self.brake_sub):
-            print('Previous subscription deleted')
-        else:
-            print('There was no previous subscription')
-        topic_path = self.brake_topic_to_subscribe.text()
-        topic_type, topic_name, fields = get_topic_type(self.node, topic_path)
-        self.brake_field_evals = generate_field_evals(fields)
-        if topic_type is not None:
-            print('Subscribing to:', topic_name, 'Type:', topic_type, 'Field:', fields)
-            data_class = get_message(topic_type)
-            self.brake_sub = self.node.create_subscription(
-                data_class,
-                topic_name,
-                self.brake_callback,
-                10)
-
-    def throttle_callback(self, msg):
-        value = msg
-        for f in self.throttle_field_evals:
-            value = f(value)
-        if value is not None and (type(value) == int or type(value) == float
-                                  or type(value) == str):
-            raw_value = int(value*100)
-            if value < 0:
-                value = 0
-                print('The throttle pedal value is not between 0 and 1')
-            elif value > 1:
-                value = 1
-                print('The throttle pedal value is not between 0 and 1')
-
-            self.updateThrottleValueSignal.emit(int(value*100), raw_value)
-        else:
-            print('The throttle pedal value is not valid')
-
-    def brake_callback(self, msg):
-        value = msg
-        for f in self.brake_field_evals:
-            value = f(value)
-        if value is not None and (type(value) == int or type(value) == float
-                                  or type(value) == str):
-            raw_value = int(value*100)
-            if value < 0:
-                value = 0
-                print('The brake pedal value is not between 0 and 1')
-            elif value > 1:
-                value = 1
-                print('The brake pedal value is not between 0 and 1')
-
-            self.updateBrakeValueSignal.emit(int(value*100), raw_value)
-        else:
-            print('The brake pedal value is not valid')
